refactor: shared selectors utils
