+++ electron/app/createWindow.ts
@@ -14,14 +14,13 @@ import {
   NEW_VERSION_CHECK_INTERVAL,
 } from '@constants/constants';
 
-import {activeProjectSelector, unsavedResourcesSelector} from '@redux/selectors';
-
 import {disableSegment, enableSegment, getSegmentClient} from '@utils/segment';
 import {StartupFlags} from '@utils/startupFlag';
 
 import {DISABLED_TELEMETRY} from '@monokle-desktop/shared/constants/telemetry';
 import {AlertEnum, AlertType, NewVersionCode} from '@monokle-desktop/shared/models';
 import utilsElectronStore from '@monokle-desktop/shared/utils/electronStore';
+import {activeProjectSelector, unsavedResourcesSelector} from '@monokle-desktop/shared/utils/selectors';
 import * as Splashscreen from '@trodi/electron-splashscreen';
 
 import autoUpdater from './autoUpdater';
