+++ electron/app/setWindowTitle.ts
@@ -2,9 +2,10 @@ import {BrowserWindow} from 'electron';
 
 import {ROOT_FILE_ENTRY} from '@constants/constants';
 
-import {isInPreviewModeSelector, kubeConfigContextSelector} from '@redux/selectors';
+import {kubeConfigContextSelector} from '@redux/selectors';
 
 import type {HelmChart, HelmValuesFile, K8sResource, RootState} from '@monokle-desktop/shared/models';
+import {isInPreviewModeSelector} from '@monokle-desktop/shared/utils/selectors';
 
 export const setWindowTitle = (state: RootState, window: BrowserWindow, projectName?: String) => {
   if (window.isDestroyed()) {
