+++ shared/package.json
@@ -26,6 +26,7 @@
     "nodemon": "2.0.20",
     "react": "18.2.0",
     "react-monaco-editor": "0.50.1",
+    "reselect": "^4.1.7",
     "runtypes": "6.6.0",
     "type-fest": "3.2.0",
     "yaml": "2.1.3"
