+++ shared/src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './electronStore';
 export * from './fileSystem';
+export * from './selectors';
