+++ src/components/organisms/PageHeader/ProjectSelection.tsx
@@ -25,7 +25,6 @@ import {openGitCloneModal} from '@redux/git';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {setCreateProject, setDeleteProject, setOpenProject, updateProjectsGitRepo} from '@redux/reducers/appConfig';
 import {openCreateProjectModal} from '@redux/reducers/ui';
-import {activeProjectSelector, unsavedResourcesSelector} from '@redux/selectors';
 
 import {Walkthrough} from '@molecules';
 
@@ -37,6 +36,7 @@ import {getRelativeDate} from '@utils';
 import {promiseFromIpcRenderer} from '@utils/promises';
 
 import {Project} from '@monokle-desktop/shared/models';
+import {activeProjectSelector, unsavedResourcesSelector} from '@monokle-desktop/shared/utils/selectors';
 
 import * as S from './ProjectSelection.styled';
