+++ src/components/organisms/SearchPane/SearchPane.tsx
@@ -25,7 +25,6 @@ import {
   setActiveTab,
   setExpandedSearchedFiles,
 } from '@redux/reducers/ui';
-import {isInPreviewModeSelector} from '@redux/selectors';
 import {getAbsoluteFilePath} from '@redux/services/fileEntry';
 
 import {TitleBar} from '@atoms';
@@ -37,6 +36,7 @@ import {filterFilesByQuery, getRegexp, notEmpty} from '@utils/filterQuery';
 import {replaceInFiles} from '@utils/replaceInFiles';
 
 import {CurrentMatch, FileEntry, MatchNode} from '@monokle-desktop/shared/models';
+import {isInPreviewModeSelector} from '@monokle-desktop/shared/utils/selectors';
 
 import TreeItem from '../FileTreePane/TreeItem';
 import {FilterTreeNode, TreeNode} from '../FileTreePane/types';
