+++ src/components/organisms/StartProjectPane/RecentProjectsPage.tsx
@@ -2,7 +2,6 @@ import {openGitCloneModal} from '@redux/git';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {setOpenProject, sortProjects, toggleProjectPin} from '@redux/reducers/appConfig';
 import {openCreateProjectModal, openFolderExplorer, toggleStartProjectPane} from '@redux/reducers/ui';
-import {activeProjectSelector} from '@redux/selectors';
 
 import SelectFolder from '@assets/FromFolder.svg';
 import CreateFromGit from '@assets/FromGit.svg';
@@ -10,6 +9,7 @@ import CreateScratch from '@assets/FromScratch.svg';
 import CreateFromTemplate from '@assets/FromTemplate.svg';
 
 import {Project} from '@monokle-desktop/shared/models';
+import {activeProjectSelector} from '@monokle-desktop/shared/utils/selectors';
 
 import Guide from './Guide';
 import RecentProject from './RecentProject';
