+++ src/components/organisms_new/SettingsPane/CurrentProjectSettings/CurrentProjectSettings.tsx
@@ -1,8 +1,9 @@
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {changeCurrentProjectName, updateProjectConfig} from '@redux/reducers/appConfig';
-import {activeProjectSelector, currentConfigSelector} from '@redux/selectors';
+import {currentConfigSelector} from '@redux/selectors';
 
 import {Project, ProjectConfig} from '@monokle-desktop/shared/models';
+import {activeProjectSelector} from '@monokle-desktop/shared/utils/selectors';
 
 import {Settings} from '../Settings/Settings';
