+++ src/hooks/resourceHooks/useDiff.ts
@@ -5,10 +5,11 @@ import {DiffTooltip, KubeConfigNoValid} from '@constants/tooltips';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {setAlert} from '@redux/reducers/alert';
 import {openResourceDiffModal} from '@redux/reducers/main';
-import {knownResourceKindsSelector, kubeConfigContextSelector, kubeConfigPathValidSelector} from '@redux/selectors';
+import {knownResourceKindsSelector, kubeConfigContextSelector} from '@redux/selectors';
 import {isKustomizationPatch, isKustomizationResource} from '@redux/services/kustomize';
 
 import {AlertEnum, AlertType, K8sResource} from '@monokle-desktop/shared/models';
+import {kubeConfigPathValidSelector} from '@monokle-desktop/shared/utils/selectors';
 
 export const useDiff = (resource?: K8sResource) => {
   const dispatch = useAppDispatch();
