+++ src/hooks/resourceHooks/useInstallDeploy.ts
@@ -4,11 +4,12 @@ import {HELM_CHART_ENTRY_FILE} from '@constants/constants';
 import {ApplyFileTooltip, ApplyTooltip, InstallValuesFileTooltip, KubeConfigNoValid} from '@constants/tooltips';
 
 import {useAppSelector} from '@redux/hooks';
-import {knownResourceKindsSelector, kubeConfigPathValidSelector} from '@redux/selectors';
+import {knownResourceKindsSelector} from '@redux/selectors';
 import {isHelmTemplateFile, isHelmValuesFile} from '@redux/services/helm';
 import {isKustomizationPatch, isKustomizationResource} from '@redux/services/kustomize';
 
 import {K8sResource} from '@monokle-desktop/shared/models';
+import {kubeConfigPathValidSelector} from '@monokle-desktop/shared/utils/selectors';
 
 export const useInstallDeploy = (resource?: K8sResource) => {
   const isKubeConfigPathValid = useAppSelector(kubeConfigPathValidSelector);
