+++ src/navsections/ImagesSectionBlueprint/ImagesSectionNameDisplay.tsx
@@ -3,9 +3,10 @@ import {useDebounce} from 'react-use';
 
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {setImagesSearchedValue} from '@redux/reducers/main';
-import {isInClusterModeSelector, isInPreviewModeSelector, kubeConfigContextColorSelector} from '@redux/selectors';
+import {isInClusterModeSelector, kubeConfigContextColorSelector} from '@redux/selectors';
 
 import {SectionInstance} from '@monokle-desktop/shared/models';
+import {isInPreviewModeSelector} from '@monokle-desktop/shared/utils/selectors';
 
 import * as S from './ImagesSectionNameDisplay.styled';
