+++ src/navsections/K8sResourceSectionBlueprint/ResourceKindSectionNameSuffix.tsx
@@ -10,12 +10,12 @@ import {ROOT_FILE_ENTRY, TOOLTIP_DELAY} from '@constants/constants';
 
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {openNewResourceWizard} from '@redux/reducers/ui';
-import {isInPreviewModeSelector} from '@redux/selectors';
 
 import {getResourceKindHandler} from '@src/kindhandlers';
 
 import {NewResourceWizardInput, SectionCustomComponentProps} from '@monokle-desktop/shared/models';
 import {Colors} from '@monokle-desktop/shared/styles/colors';
+import {isInPreviewModeSelector} from '@monokle-desktop/shared/utils/selectors';
 
 const SuffixContainer = styled.span`
   display: inline-block;
